refactor: lets move components to their own dir
