+++ src/components/App/App.tsx
@@ -1,14 +1,31 @@
 import * as React from 'react';
 import './App.css';
 
-interface NameProps {
-  name?: string;
-}
+// PropTypes can be setup as Interfaces
+// interface NameProps {
+//   name?: string;
+// }
 
-interface LogoProps {
+// interface LogoProps {
+//   url?: string;
+// }
+
+// `type` seems to be better for props and state however
+
+type LogoProps = {
   url?: string;
-}
+};
 
+type NameProps = {
+  name?: string;
+};
+
+
+/*
+ url and name are given a default prop here as a default function parameter
+ instead of defining `defaultProps`
+ https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/Default_parameters
+*/
 const Logo: React.SFC<LogoProps> = ({ url = '/logo512.png' }) => (
   <img className="App-logo" alt="logo" src={url} />
 );
@@ -19,7 +36,7 @@ const Name: React.SFC<NameProps> = ({ name = 'React Developer' }) => (
   </p>
 );
 
-export const App: React.SFC = () => (
+const App: React.SFC = () => (
   <div className="App">
     <header className="App-header">
       <Logo />
